Fixed initial tests for usePolling hook
